Extract Base functionality from S3 LoadStrategy
Authors: @darnould, @integralist

Closes #38
