+++ alephant-broker.gemspec
@@ -33,7 +33,7 @@ Gem::Specification.new do |spec|
 
   spec.add_runtime_dependency "alephant-lookup"
   spec.add_runtime_dependency "alephant-cache"
-  spec.add_runtime_dependency 'alephant-logger'
+  spec.add_runtime_dependency "alephant-logger", "~> 0"
   spec.add_runtime_dependency 'alephant-sequencer'
   spec.add_runtime_dependency "dalli-elasticache"
   spec.add_runtime_dependency "pmap"
